Reorganize vendor hash modules
