+++ .istanbul.yml
@@ -1,4 +1,3 @@
 instrumentation:
   excludes:
-  - 'lib/base62-encode.js'
-  - 'lib/hash-string.js'
+  - 'dist/hash-utils/**'
