+++ src/hash-utils/hash-string.js
